Remove ClientViewElements and enhance Management Feature
This commit removes the `ClientViewElements` from the code as the necessary functions have been moved directly into the `MainLayout`. This clean-up and enhancement to the `MainLayout` class promotes better code organization and readability.

Specifically, the `newClientButton` function is moved to the `MainLayout`. Enhanced client management functionalities which include the ability to drag and drop to rearrange client order and the ability to add a new client directly from the navigation menu have been added. These changes improve the user's ability to manage clients more dynamically.

Furthermore, all client-related operations are now handled in `ClientService` itself removing the need for multiple components dealing with client operations.

In terms of UI, the new update introduces new styles for displaying client items and grab handles making the drag and drop functionality more explicit to the end-user.

Also, the `ClientView` class is introduced that handles the parameterized routes functionality where routing and redirection to views based on client ID are done.

The changes in this commit enhance the application's user interface and experience by providing a more intuitive and user-friendly way of managing clients.