Update ClientService for optimal transaction management
This commit adds `@Transactional` annotation to `ClientService` class to ensure optimal and consistent transactions. This addition improves the reliability of operations that depend on multiple database actions.

Remove redundant UI components in MainLayout

The redundancy of adding empty space and creating a button was removed from `MainLayout` class by minimizing methods for a cleaner layout. Also, the `MainLayout` class now implements `DesignTools` to conform to the single responsibility principle improving code maintainability.

Create DesignTools for UI space management

Introduced a new interface `DesignTools` to manage adding empty space in the layout. This addition centralizes the component creation in one place which makes the code organization cleaner and reduces code duplication.

Update MainLayout with new CSS styles

Added new CSS styles `.my-text-field` and `.italic` to emphasize text in the application's interface which enhances the visual presentation of the text fields.

Refactor ClientView

This commit refactors `ClientView` class and optimizes the way client information is handled and displayed. It modifies the setter method for client parameters to take care of all client related updates ensuring more precise data handling.

This also introduces checks for user roles and permissions to view and edit the client's name. A user with the 'ADMIN' role is granted the ability to update a client's name improving the functionality of client management for admin users.

The UI display was additionally refactored to improve readability and the display of information about the client. It offers a more structured way of displaying client information providing an enhanced user experience.