Enhance UI and improve Note sorting functionality
Updated the user interface by modifying the loadNotes method in NoteGridService and including a pin icon in NoteCard. Improvements were also made on role-based operations by adding a user parameter in loadNotes. To maintain consistency, the method name in AddNewNoteService was refactored from createNewProductDialog to createNewNoteDialog. The UI design was tweaked via CSS modifications to the 'smaller-text' class. Functionality for sorting notes by 'pinned' status and creation time was added to NoteRepository.