Update design tools and components, add edit functionality for notes
Modifications have been made to the design tools class to enhance the creation of text areas and combo boxes. The 'NoteCard' class has been enhanced with the addition of an edit option. This involves the introduction of the 'NoteDialog' class which encapsulates the editing process of a note. Several classes have also been equipped with the new 'Noteable' interface to support this revised workflow structure efficiently.