Add product shelf life, fetch orders by product, show orders in ProductView
The code adds a "shelfLife" field to products and updates the functionality for adding a new product to include this new field. It also implements the option to retrieve all orders associated with a specific product and visualizes these orders in the ProductView. The code changes mainly improve the overview and management of products, as well as their associated orders.