Add notes button to user orders and filtering by order status
Updates were made to the OrdersGrid.java and OrderComponents.java to accommodate new features. Users are now able to add notes to their orders with the addition of a notes button that opens a dialog. Furthermore, status filtering is now available on the order page to enhance user experience. Minor tweaks were made to the database connection in the application.properties file.